Implemented new sprite data structure
This is the first step towards separating entities from their sprites
Going forward entities will be code attached to a sprite.

This will allow to implement displaylists streamlining the display of
sprites to the screen albeit from an entity or a standalone sprite.

This will make the designed to general purpose elements such as huds
easier.