Added the OnDraw() event to both Scene & Entity
Added the OnDraw() event to allow game specific custom rendering once
the engine's done drawing the screen.

This will allow game to be more creative regarding their rendering
routines

This was the ocasion de reorder said events to their order matches
between Scenes and entity

Added KON_LoadSpriteKeyed() as a helper function to KON_LoadSprite() to
reduce the dependencies to Load.h of games using it.