Added the ability to toogle integer scalling
Koneko now allows the user to toogle integer scalling, either setting
the integerScalling flag of DisplayDevice or by hitting the F2 key.

Note that the integer scalling status is ignored for resolution beneath
the internal resolution of the game as there's nothing to gain from it
in such a case.

Also, integer scalling is now dissabled by default as a large majority
of games don't benifit from it.