Surface: Added new KON_Surface type to better abstract from SDL
- Reworked the engine so it displays KON_Surface(s) instead of
SDL_Texture(s).

Thanks to this change the user is forced to use Koneko's load/free and
displays functions so to benefit from the engine's advantage and not to
break the ressource manager.

The less coupled to the SDL Koneko is, the easier will it be to port to
another library if need be. This will notably allow Koneko to be ported
on devices where SDL is abscent.

- Moved Draw funtions from System to Surface, as we don't want to expose
  any SDL type to the user through headers.