Raycast: Fixed dda precision
Previously the rays were being casted from the perspective of the
camera. This in turned caused innacuracy in the DDA due to floating
point rounding errors.

This has now been changed to be casted from the perspective of the layer
instead. The DDA also iterates on unsigned int which in turned makes
this kind of innacuracies impossible.