DataTypes: Removed Referenced Linked List operations
Adding/Inserting by reference causes incertainties relative to who
should free the data in a LinkedList.

This was previously added as an optimisation to save on memcpys
but it would be fool to fall in an early optimisation trap this
early in the lifetime of the project.

This was removed in favor of copy on append LinkedList.