Switch: Added switch build target !
It is now possible to build Koneko targeting the switch.
When targetting the switch Koneko's dependencies will be:
- SDL2 + mixer
- libXml2

As a bonus it is also possible to build the linux version with libXml2
as XML backend instead of mxml in order to test against the same
configuration as on the switch.